feat: replace reactFlow start base to FlowBoard component
