+++ src/Main/Main.css
@@ -1,4 +1,6 @@
 .main {
   /* height: 900px; */
   background-color: #F1C88B;
+  width: 100vw;
+  height: 100vh;
 }
